requirements.txt
influxdb
jsonrpcclient<3
request
pytz